Added room_xml_parser.hpp and moved implementation out of the header file
